requirements.txt
aiogram>=3.15
sqlalchemy[asyncio]>=2.0.36
asyncpg
aiosqlite
python-dotenv
alembic
apscheduler>=3.10
